requirements.txt
discord.py==1.6.*
environs==9.3.*
psycopg2-binary==2.9.*
SQLAlchemy==1.4.*